Consolidating types in in a single file per scope (dbase/business)
