+++ sources/types/businesslogic/business.types.ts
@@ -1,0 +1,19 @@
+﻿import { GuildChannel } from 'discord.js';
+import { WatchedPlayer } from './../dbase/persisted.types';
+
+export interface MappedGuildConfiguration {
+    id: string,
+    defaultChannel: GuildChannel,
+    emergencyChannel: GuildChannel
+}
+
+export interface ScannedFaction {
+    name: string;
+    playersCount: number;
+}
+
+export interface ScannedRegion {
+    name: string;
+    watchedPlayers: WatchedPlayer[];
+    playersCount: number;
+}
