--- sources/types/businesslogic/mapped.guild.configuration.type.ts
@@ -1,7 +1,0 @@
-﻿import { GuildChannel } from 'discord.js';
-
-export interface MappedGuildConfiguration {
-    id: string,
-    defaultChannel: GuildChannel,
-    emergencyChannel: GuildChannel
-}
