--- sources/types/businesslogic/scanned.faction.type.ts
@@ -1,13 +1,0 @@
-﻿export interface ScannedFaction {
-    name: string;
-    playersCount: number;
-}
-
-export function isScannedFaction(obj: any): obj is ScannedFaction {
-    if (typeof obj.name === "string" &&
-        typeof obj.playersCount === "number") {
-        return true;
-    }
-
-    return false;
-}
