--- sources/types/dbase/business/activity.cache.item.type.ts
@@ -1,11 +1,0 @@
-﻿export interface ActivityCacheItem {
-    name: string;
-    playersCount: number;
-}
-
-export function isActivityCacheItem(obj: any): obj is ActivityCacheItem {
-    return (
-        typeof obj.name === "string" &&
-        typeof obj.playersCount === "number"
-    );
-}
